+++ podcasts/PodcastChapterParser.swift
@@ -26,10 +26,9 @@ class PodcastChapterParser {
                     guard let chapters = MNAVChapterReader.chapters(from: movieAsset) as? [MNAVChapter], chapters.count > 0 else { return }
 
                     var parsedChapters = [ChapterInfo]()
-                    for (index, chapter) in chapters.enumerated() {
+                    var index = 0
+                    for chapter in chapters {
                         let convertedChapter = ChapterInfo()
-                        convertedChapter.isFirst = index == 0
-                        convertedChapter.isLast = (index == chapters.count - 1)
                         convertedChapter.title = chapter.title ?? ""
                         #if !os(watchOS)
                             convertedChapter.image = chapter.artwork
@@ -37,14 +36,19 @@ class PodcastChapterParser {
 
                         convertedChapter.startTime = chapter.time
                         convertedChapter.duration = chapter.duration.seconds
-                        convertedChapter.index = index
+                        convertedChapter.isHidden = chapter.hidden
+                        if !convertedChapter.isHidden {
+                            convertedChapter.index = index
+                            index += 1
+                        }
                         if strongSelf.isValidUrl(chapter.url) {
                             convertedChapter.url = chapter.url
                         }
 
                         parsedChapters.append(convertedChapter)
                     }
-
+                    parsedChapters.first(where: {!$0.isHidden})?.isFirst = true
+                    parsedChapters.last(where: {!$0.isHidden})?.isLast = true
                     completion(parsedChapters)
                 }
             } catch {